+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.0.8] - 2018-05-24
+### Added
+- Broker order queue and unit tests
+- Exchange websocket feeds collection with loader and closer with tests
+
 ## [0.0.7] - 2018-03-31
 ### Added
 - Mock websocket events for heartbeat and ticker messages